Move matchesEngines to a method on Environment objects
