+++ packages/core/utils/package.json
@@ -17,7 +17,6 @@
   "dependencies": {
     "@iarna/toml": "^2.2.0",
     "@parcel/logger": "^2.0.0-alpha.1.1",
-    "browserslist": "^4.6.6",
     "clone": "^2.1.1",
     "deasync": "^0.1.14",
     "fast-glob": "^2.2.2",
@@ -27,7 +26,6 @@
     "node-forge": "^0.8.1",
     "nullthrows": "^1.1.1",
     "resolve": "^1.12.0",
-    "semver": "^5.4.1",
     "serialize-to-js": "^1.1.1",
     "terser": "^3.7.3"
   },
